Make the exported node tests run on PHP 7.4
Three ways the new tests did not survive the 7.4 job:

- The name scope collecting visitor was an anonymous class taking the result array as a
  by-reference promoted property. simple-downgrade expands promotion into a plain assignment,
  which drops the reference, so the caller kept its empty array. It is a named test class now
  and hands the scopes back through a getter.
- The property hooks in exported-properties-hooks.php do not parse below 8.4 - the parser
  follows the analysed PHP version, which defaults to the running one - so the test that needs
  them is marked #[RequiresPhp('>= 8.4.0')].
- That data file is not in the lint exclude list either, so its first line carries `// lint >= 8.4`.

Verified against a downgraded tree on a real php 7.4 binary, not just the php-parser grammar.
