+++ tests/PHPStan/Dependency/data/exported-properties-hooks.php
@@ -1,4 +1,4 @@
-<?php declare(strict_types = 1);
+<?php declare(strict_types = 1); // lint >= 8.4
 
 namespace ExportedPropertiesHooks;
